(matchers) Node bindings tree builder
The bound_nodes_tree_builder class is passed as part of the matcher
interface, its purpose is to collect all named matching nodes in a
matching subtree. The builder constructs a hierarchical record of the
matching nodes found and bound to an specific id string. The structure
of the matching ASt is preserved by this hierarchical recording of
nodes. To obtain the full set of found bound nodes, the builder provides
a visitor interface that returns a set of bound nodes for each match
found in the AST.